displayer/requirements.txt
joblib==1.4.2
numpy==2.1.1
opencv-python==4.10.0.84
paho-mqtt==2.1.0
pandas==2.2.3
pillow==10.4.0
protobuf==3.20.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.2
